Fix some build errors

+++ src/cradle/caching/immutable_cache.cpp
@@ -1,5 +1,6 @@
 #include <cradle/caching/immutable_cache.h>
 
+#include <atomic>
 #include <list>
 #include <mutex>
 #include <unordered_map>
@@ -151,7 +152,8 @@ reduce_memory_cache_size(immutable_cache& cache, int desired_size)
     {
         std::scoped_lock<std::mutex> lock(cache.mutex);
         while (!cache.eviction_list.records.empty()
-               && cache.eviction_list.total_size > desired_size * 0x100000)
+               && cache.eviction_list.total_size
+                      > size_t(desired_size) * 0x100000)
         {
             auto const& i = cache.eviction_list.records.front();
             auto data_size = i->data.ptr ? i->data.ptr->deep_size() : 0;
